Add domain argument to deleteCookies
This fixes an issue where we were unable to delete GA cookies because
of they were set on a different domain. GA cookies are set to match all
subdomains i.e. ".www.civilservice.lgbt", however we were only to delete
cookies for the exact domain "www.civilservice.lgbt". This adds a domain
argument to be able to specify the domain in which to set and delete
cookies from.